requirements.txt
# GUI Framework
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
PyQt6-sip>=13.6.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.13.0
# Supabase for cloud sync
supabase>=2.0.0
# Optional: For PostgreSQL support (uncomment if needed)
# psycopg2-binary>=2.9.9
# Optional: For MySQL support (uncomment if needed)
# pymysql>=1.1.0

# AI/ML
# Option 1: Ollama (recommended - smaller, faster)
ollama>=0.1.0
# Option 2: Hugging Face (alternative)
# transformers>=4.35.0
# torch>=2.1.0
# sentencepiece>=0.1.99
# accelerate>=0.25.0
# huggingface_hub>=0.19.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0
python-dateutil>=2.8.2

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Utilities
python-dotenv>=1.0.0
bcrypt>=4.1.0
cryptography>=41.0.0
requests>=2.31.0
packaging>=23.0

# Charts/Visualization
pyqtgraph>=0.13.0
matplotlib>=3.8.0

# Testing
pytest>=7.4.0
pytest-qt>=4.2.0

# Optional: Report Generation
reportlab>=4.0.0

# ML/Analytics
scikit-learn>=1.3.0
joblib>=1.3.0

# Document Processing
pypdf>=3.17.0
pdfminer.six>=20221105

# Email Integration
smtplib  # Built-in
email-validator>=2.0.0

# Payment Integration (optional)
# stripe>=7.0.0
# paypalrestsdk>=1.13.0

# Caching & Performance
cachetools>=5.3.0
aiohttp>=3.9.0

# Security
keyring>=24.0.0

